[android] Refactor external link settings screen
