+++ app_manager/appmanager/src/main/java/app/editors/manager/ui/fragments/share/link/ExternalLinkBlock.kt
@@ -14,33 +14,136 @@ import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material.Icon
 import androidx.compose.material.IconButton
 import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Surface
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.colorResource
 import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import app.documents.core.network.share.models.ExternalLink
 import lib.compose.ui.addIfNotNull
 import lib.compose.ui.theme.ManagerTheme
 import lib.compose.ui.theme.colorTextTertiary
-import lib.compose.ui.views.AppScaffold
+import lib.compose.ui.views.AppDescriptionItem
+import lib.compose.ui.views.AppHeaderItem
+import lib.compose.ui.views.AppTextButton
 import lib.toolkit.base.R
+import lib.toolkit.base.managers.utils.openSendTextActivity
 
 
 @Composable
-fun ExternalLinkItem(
+internal fun ExternalLinkBlock(
+    generalLink: ExternalLink?,
+    additionalLinks: List<ExternalLink>,
+    canEditRoom: Boolean,
+    onLinkClick: (ExternalLink) -> Unit,
+    onGeneralLinkCreate: () -> Unit,
+    onAdditionalLinkCreate: () -> Unit,
+    onCopyLinkClick: (String) -> Unit
+) {
+    val context = LocalContext.current
+    if (!canEditRoom && (generalLink != null || additionalLinks.isNotEmpty()) || canEditRoom) {
+        AppDescriptionItem(
+            modifier = Modifier.padding(top = 8.dp),
+            text = app.editors.manager.R.string.rooms_info_access_desc
+        )
+    }
+    if (!canEditRoom && generalLink != null || canEditRoom) {
+        AppHeaderItem(title = app.editors.manager.R.string.rooms_info_general_link)
+    }
+    if (generalLink != null) {
+        ExternalLinkItem(
+            linkTitle = generalLink.sharedTo.title,
+            access = generalLink.access,
+            hasPassword = !generalLink.sharedTo.password.isNullOrEmpty(),
+            expiring = false,
+            isExpired = generalLink.sharedTo.isExpired,
+            canEdit = canEditRoom,
+            onCopyLinkClick = { onCopyLinkClick.invoke(generalLink.sharedTo.shareLink) },
+            onShareClick = {
+                context.openSendTextActivity(
+                    context.getString(app.editors.manager.R.string.toolbar_menu_main_share),
+                    generalLink.sharedTo.shareLink
+                )
+            },
+            onClick = { onLinkClick.invoke(generalLink) }.takeIf { canEditRoom }
+        )
+    }
+
+    if (canEditRoom && generalLink == null) {
+        AppTextButton(
+            modifier = Modifier.padding(start = 8.dp),
+            title = app.editors.manager.R.string.rooms_info_create_link,
+            onClick = onGeneralLinkCreate
+        )
+    }
+
+    if (!canEditRoom && additionalLinks.isNotEmpty() || canEditRoom) {
+        Row {
+            AppHeaderItem(
+                modifier = Modifier.weight(1f),
+                title = stringResource(
+                    id = app.editors.manager.R.string.rooms_info_additional_links,
+                    additionalLinks.size,
+                    RoomInfoFragment.MAX_ADDITIONAL_LINKS_COUNT
+                )
+            )
+            if (additionalLinks.size in 1..4) {
+                IconButton(onClick = onAdditionalLinkCreate) {
+                    Icon(
+                        imageVector = ImageVector.vectorResource(app.editors.manager.R.drawable.ic_action_button_docs_add),
+                        tint = MaterialTheme.colors.primary,
+                        contentDescription = null
+                    )
+                }
+            }
+        }
+    }
+
+    additionalLinks.forEach { link ->
+        ExternalLinkItem(
+            linkTitle = link.sharedTo.title,
+            access = link.access,
+            hasPassword = !link.sharedTo.password.isNullOrEmpty(),
+            expiring = !link.sharedTo.expirationDate.isNullOrEmpty(),
+            isExpired = link.sharedTo.isExpired,
+            canEdit = canEditRoom,
+            onCopyLinkClick = { onCopyLinkClick.invoke(link.sharedTo.shareLink) },
+            onShareClick = {
+                context.openSendTextActivity(
+                    context.getString(app.editors.manager.R.string.toolbar_menu_main_share),
+                    link.sharedTo.shareLink
+                )
+            },
+            onClick = { onLinkClick.invoke(link) }.takeIf { canEditRoom }
+        )
+    }
+    if (additionalLinks.isEmpty() && canEditRoom) {
+        AppTextButton(
+            modifier = Modifier.padding(start = 8.dp),
+            title = app.editors.manager.R.string.rooms_info_create_link,
+            onClick = onAdditionalLinkCreate
+        )
+    }
+}
+
+@Composable
+private fun ExternalLinkItem(
     linkTitle: String,
     access: Int,
     hasPassword: Boolean,
     expiring: Boolean,
     isExpired: Boolean,
     canEdit: Boolean,
+    onCopyLinkClick: () -> Unit,
     onShareClick: () -> Unit,
     onClick: (() -> Unit)? = null
 ) {
@@ -85,7 +188,14 @@ fun ExternalLinkItem(
             }
         }
         if (canEdit) {
-            IconButton(modifier = Modifier.padding(end = 24.dp), onClick = onShareClick) {
+            IconButton(modifier = Modifier, onClick = onCopyLinkClick) {
+                Icon(
+                    imageVector = ImageVector.vectorResource(app.editors.manager.R.drawable.ic_list_context_external_link),
+                    tint = MaterialTheme.colors.primary,
+                    contentDescription = null
+                )
+            }
+            IconButton(modifier = Modifier, onClick = onShareClick) {
                 Icon(
                     imageVector = ImageVector.vectorResource(app.editors.manager.R.drawable.ic_list_context_share),
                     tint = MaterialTheme.colors.primary,
@@ -93,7 +203,7 @@ fun ExternalLinkItem(
                 )
             }
             Icon(
-                modifier = Modifier.padding(end = 12.dp),
+                modifier = Modifier.padding(horizontal = 8.dp),
                 imageVector = ImageVector.vectorResource(R.drawable.ic_arrow_right),
                 tint = MaterialTheme.colors.colorTextTertiary,
                 contentDescription = null
@@ -133,7 +243,7 @@ private fun AttributeIcon(icon: Int, activate: Boolean = true) {
 @Composable
 private fun Preview() {
     ManagerTheme {
-        AppScaffold {
+        Surface {
             Column {
                 ExternalLinkItem(
                     linkTitle = "Shared link",
@@ -142,6 +252,7 @@ private fun Preview() {
                     expiring = false,
                     isExpired = false,
                     canEdit = true,
+                    {},
                     {}
                 ) {}
                 ExternalLinkItem(
@@ -150,6 +261,7 @@ private fun Preview() {
                     expiring = false,
                     isExpired = true,
                     canEdit = false,
+                    {},
                     {}
                 ) {}
             }
