+++ app_manager/appmanager/src/main/java/app/editors/manager/ui/fragments/share/link/RoomInfoFragment.kt
@@ -9,21 +9,12 @@ import androidx.activity.ComponentDialog
 import androidx.activity.compose.BackHandler
 import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.result.contract.ActivityResultContracts
-import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.verticalScroll
-import androidx.compose.material.Icon
-import androidx.compose.material.IconButton
+import androidx.compose.material.CircularProgressIndicator
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Surface
 import androidx.compose.material.Text
@@ -31,21 +22,14 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
+import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.draw.rotate
-import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.platform.ComposeView
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.platform.LocalView
+import androidx.compose.ui.platform.LocalSoftwareKeyboardController
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavType
 import androidx.navigation.compose.NavHost
@@ -68,27 +52,17 @@ import app.editors.manager.ui.dialogs.fragments.BaseDialogFragment
 import app.editors.manager.viewModels.link.RoomInfoEffect
 import app.editors.manager.viewModels.link.RoomInfoState
 import app.editors.manager.viewModels.link.RoomInfoViewModel
-import com.bumptech.glide.integration.compose.ExperimentalGlideComposeApi
-import com.bumptech.glide.integration.compose.GlideImage
-import com.bumptech.glide.integration.compose.placeholder
+import kotlinx.coroutines.delay
 import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
-import lib.compose.ui.addIf
+import lib.compose.ui.rememberWaitingDialog
 import lib.compose.ui.theme.ManagerTheme
-import lib.compose.ui.theme.colorTextSecondary
-import lib.compose.ui.theme.colorTextTertiary
-import lib.compose.ui.views.AnimatedVisibilityVerticalFade
-import lib.compose.ui.views.AppDescriptionItem
-import lib.compose.ui.views.AppDivider
-import lib.compose.ui.views.AppHeaderItem
 import lib.compose.ui.views.AppScaffold
-import lib.compose.ui.views.AppTextButton
 import lib.compose.ui.views.AppTopBar
 import lib.compose.ui.views.TopAppBarAction
 import lib.toolkit.base.managers.utils.KeyboardUtils
 import lib.toolkit.base.managers.utils.UiUtils
 import lib.toolkit.base.managers.utils.getSerializableExt
-import lib.toolkit.base.managers.utils.openSendTextActivity
 import lib.toolkit.base.managers.utils.putArgs
 
 class RoomInfoFragment : BaseDialogFragment() {
@@ -96,7 +70,7 @@ class RoomInfoFragment : BaseDialogFragment() {
     companion object {
 
         private const val KEY_ROOM = "key_room"
-        private const val MAX_ADDITIONAL_LINKS_COUNT = 5
+        internal const val MAX_ADDITIONAL_LINKS_COUNT = 5
         val TAG: String = RoomInfoFragment::class.java.simpleName
 
         fun newInstance(room: CloudFolder): RoomInfoFragment =
@@ -119,12 +93,12 @@ class RoomInfoFragment : BaseDialogFragment() {
         return ComposeView(requireContext())
     }
 
+    @OptIn(ExperimentalComposeUiApi::class)
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         (view as? ComposeView)?.setContent {
             ManagerTheme {
-                val context = LocalContext.current
-                val localView = LocalView.current
-                val portal = remember { context.accountOnline?.portal }
+                val keyboardController = LocalSoftwareKeyboardController.current
+                val portal = remember { requireContext().accountOnline?.portal }
                 val room = remember { arguments?.getSerializableExt<CloudFolder>(KEY_ROOM) }
                 val canEditRoom = room?.security?.editRoom == true
                 val viewModel = viewModel { RoomInfoViewModel(requireContext().roomProvider, room?.id.orEmpty()) }
@@ -137,25 +111,32 @@ class RoomInfoFragment : BaseDialogFragment() {
                 }
                 val state by viewModel.state.collectAsState()
                 val shareActivityResult = rememberLauncherForActivityResult(
-                    ActivityResultContracts.StartActivityForResult()
-                ) {
-                    viewModel.fetchRoomInfo()
-                }
+                    contract = ActivityResultContracts.StartActivityForResult(),
+                    onResult = { viewModel.fetchRoomInfo() }
+                )
+                val waitingDialog = rememberWaitingDialog(
+                    title = R.string.dialogs_wait_title,
+                    onCancel = viewModel::cancelOperation
+                )
 
                 LaunchedEffect(Unit) {
                     viewModel.effect.collect { effect ->
                         when (effect) {
+                            is RoomInfoEffect.Create -> {
+                                copyLinkToClipboard(requireView(), effect.url, true)
+                                waitingDialog.dismiss()
+                            }
                             is RoomInfoEffect.Error -> {
                                 UiUtils.getShortSnackBar(requireView())
                                     .setText(effect.message)
                                     .show()
                             }
-                            is RoomInfoEffect.Create -> {
-                                KeyboardUtils.setDataToClipboard(context, effect.url)
-                                UiUtils.getSnackBar(localView)
-                                    .setText(R.string.rooms_info_copy_link_to_clipboard)
-                                    .show()
+                            RoomInfoEffect.ShowOperationDialog -> {
+                                keyboardController?.hide()
+                                waitingDialog.show()
+                                delay(500)
                             }
+                            RoomInfoEffect.CloseDialog -> waitingDialog.dismiss()
                         }
                     }
                 }
@@ -193,7 +174,7 @@ class RoomInfoFragment : BaseDialogFragment() {
                                 onLinkClick = { link ->
                                     val json = Json.encodeToString(link.sharedTo)
                                     navController.navigate("${RoomInfoScreens.LinkSettings.name}?link=$json")
-                                }
+                                },
                             )
                         }
                         composable(
@@ -266,10 +247,7 @@ class RoomInfoFragment : BaseDialogFragment() {
             modifier = Modifier.fillMaxSize(),
             contentAlignment = Alignment.Center
         ) {
-            Text(
-                text = stringResource(id = R.string.placeholder_loading),
-                style = MaterialTheme.typography.body1
-            )
+            CircularProgressIndicator()
         }
     }
 
@@ -327,7 +305,8 @@ class RoomInfoFragment : BaseDialogFragment() {
                             canEditRoom = canEditRoom,
                             onLinkClick = onLinkClick,
                             onGeneralLinkCreate = onGeneralLinkCreate,
-                            onAdditionalLinkCreate = onAdditionalLinkCreate
+                            onAdditionalLinkCreate = onAdditionalLinkCreate,
+                            onCopyLinkClick = { url -> copyLinkToClipboard(requireView(), url, false) }
                         )
                     }
                     ShareUsersList(
@@ -353,179 +332,18 @@ class RoomInfoFragment : BaseDialogFragment() {
         }
     }
 
-    @Composable
-    private fun ExternalLinkBlock(
-        generalLink: ExternalLink?,
-        additionalLinks: List<ExternalLink>,
-        canEditRoom: Boolean,
-        onLinkClick: (ExternalLink) -> Unit,
-        onGeneralLinkCreate: () -> Unit,
-        onAdditionalLinkCreate: () -> Unit
-    ) {
-        val context = LocalContext.current
-        if (!canEditRoom && (generalLink != null || additionalLinks.isNotEmpty()) || canEditRoom) {
-            AppDescriptionItem(
-                modifier = Modifier.padding(top = 8.dp),
-                text = R.string.rooms_info_access_desc
-            )
-        }
-        if (!canEditRoom && generalLink != null || canEditRoom) {
-            AppHeaderItem(title = R.string.rooms_info_general_link)
-        }
-        if (generalLink != null) {
-            ExternalLinkItem(
-                linkTitle = generalLink.sharedTo.title,
-                access = generalLink.access,
-                hasPassword = !generalLink.sharedTo.password.isNullOrEmpty(),
-                expiring = false,
-                isExpired = generalLink.sharedTo.isExpired,
-                canEdit = canEditRoom,
-                onShareClick = {
-                    context.openSendTextActivity(
-                        context.getString(R.string.toolbar_menu_main_share),
-                        generalLink.sharedTo.shareLink
-                    )
-                },
-                onClick = { onLinkClick.invoke(generalLink) }.takeIf { canEditRoom }
-            )
-        }
-
-        if (canEditRoom && generalLink == null) {
-            AppTextButton(
-                modifier = Modifier.padding(start = 8.dp),
-                title = R.string.rooms_info_create_link,
-                onClick = onGeneralLinkCreate
-            )
-        }
-
-        if (!canEditRoom && additionalLinks.isNotEmpty() || canEditRoom) {
-            AppHeaderItem(
-                title = stringResource(
-                    id = R.string.rooms_info_additional_links,
-                    additionalLinks.size,
-                    MAX_ADDITIONAL_LINKS_COUNT
-                )
-            )
-        }
-
-        additionalLinks.forEach { link ->
-            ExternalLinkItem(
-                linkTitle = link.sharedTo.title,
-                access = link.access,
-                hasPassword = !link.sharedTo.password.isNullOrEmpty(),
-                expiring = !link.sharedTo.expirationDate.isNullOrEmpty(),
-                isExpired = link.sharedTo.isExpired,
-                canEdit = canEditRoom,
-                onShareClick = {
-                    context.openSendTextActivity(
-                        context.getString(R.string.toolbar_menu_main_share),
-                        link.sharedTo.shareLink
-                    )
-                },
-                onClick = { onLinkClick.invoke(link) }.takeIf { canEditRoom }
-            )
-        }
-        if (additionalLinks.size < MAX_ADDITIONAL_LINKS_COUNT && canEditRoom) {
-            AppTextButton(
-                modifier = Modifier.padding(start = 8.dp),
-                title = R.string.rooms_info_create_link,
-                onClick = onAdditionalLinkCreate
-            )
-        }
-    }
-
-    @Composable
-    private fun ShareUsersList(title: Int, portal: String?, shareList: List<Share>?, onClick: (String, Int, Boolean) -> Unit) {
-        var visible by remember { mutableStateOf(true) }
-
-        if (!shareList.isNullOrEmpty()) {
-            Column {
-                Row(modifier = Modifier.padding(start = 16.dp, end = 16.dp, top = 16.dp, bottom = 8.dp)) {
-                    Text(
-                        modifier = Modifier
-                            .weight(1f)
-                            .align(Alignment.Bottom),
-                        text = stringResource(id = title, shareList.size),
-                        style = MaterialTheme.typography.subtitle2,
-                        color = MaterialTheme.colors.colorTextSecondary
-                    )
-                    IconButton(
-                        modifier = Modifier.size(24.dp),
-                        onClick = { visible = !visible }
-                    ) {
-                        Icon(
-                            modifier = Modifier.rotate(if (visible) 0f else 180f),
-                            imageVector = ImageVector.vectorResource(lib.toolkit.base.R.drawable.ic_arrow_down),
-                            contentDescription = null,
-                            tint = MaterialTheme.colors.colorTextTertiary
-                        )
-                    }
-                }
-                AnimatedVisibilityVerticalFade(visible = visible) {
-                    Column {
-                        shareList.forEach { share ->
-                            ShareUserItem(share = share, portal = portal) {
-                                onClick.invoke(
-                                    share.sharedTo.id,
-                                    share.intAccess,
-                                    share.sharedTo.isOwner || share.sharedTo.isAdmin
-                                )
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    @OptIn(ExperimentalGlideComposeApi::class)
-    @Composable
-    private fun ShareUserItem(share: Share, portal: String?, onClick: () -> Unit) {
-        Column(modifier = Modifier.addIf(share.canEditAccess) { clickable(onClick = onClick) }) {
-            Row(
-                modifier = Modifier
-                    .padding(horizontal = 16.dp)
-                    .fillMaxWidth()
-                    .height(56.dp),
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                GlideImage(
-                    modifier = Modifier
-                        .padding(end = 16.dp)
-                        .clip(CircleShape)
-                        .size(40.dp)
-                        .background(MaterialTheme.colors.colorTextTertiary),
-                    model = "${ApiContract.SCHEME_HTTPS}$portal${share.sharedTo.avatarMedium}",
-                    loading = placeholder(R.drawable.ic_account_placeholder),
-                    contentDescription = null
-                )
-                Text(
-                    modifier = Modifier.weight(1f),
-                    style = MaterialTheme.typography.body1,
-                    text = share.sharedTo.displayName
-                )
-                Text(
-                    text = stringResource(id = RoomUtils.getAccessTitleOrOwner(share)),
-                    style = MaterialTheme.typography.body2,
-                    color = if (share.canEditAccess)
-                        MaterialTheme.colors.colorTextSecondary else
-                        MaterialTheme.colors.colorTextTertiary
-                )
-                if (share.canEditAccess) {
-                    Icon(
-                        imageVector = ImageVector.vectorResource(lib.toolkit.base.R.drawable.ic_arrow_right),
-                        contentDescription = null,
-                        tint = MaterialTheme.colors.colorTextTertiary
-                    )
-                }
-            }
-            AppDivider(startIndent = 16.dp + 40.dp + 16.dp)
-        }
+    private fun copyLinkToClipboard(rootView: View, url: String, isCreate: Boolean) {
+        KeyboardUtils.setDataToClipboard(requireContext(), url)
+        UiUtils.getSnackBar(rootView).setText(
+            if (!isCreate)
+                R.string.rooms_info_copy_link_to_clipboard else
+                R.string.rooms_info_create_link_complete
+        ).show()
     }
 
     @Preview
     @Composable
-    private fun ExternalLinkScreenPreview() {
+    private fun RoomInfoScreenPreview() {
         val link = ExternalLink(
             access = 2,
             isLocked = false,
@@ -555,9 +373,9 @@ class RoomInfoFragment : BaseDialogFragment() {
                     isLoading = false,
                     generalLink = link,
                     additionalLinks = listOf(
-                        link.copy(sharedTo = link.sharedTo.copy(title = "Link 1", expirationDate = "123")),
-                        link.copy(sharedTo = link.sharedTo.copy(title = "Link 2", password = "123")),
-                        link.copy(sharedTo = link.sharedTo.copy(title = "Link 3", isExpired = true)),
+                        link.copy(sharedTo = link.sharedTo.copy(title = "Additional link 1", expirationDate = "123")),
+                        link.copy(sharedTo = link.sharedTo.copy(title = "Additional link 2", password = "123")),
+                        link.copy(sharedTo = link.sharedTo.copy(title = "Additional link 3", isExpired = true)),
                     ),
                     shareList = listOf(
                         Share(access = "1", sharedTo = SharedTo(displayName = "User 1"), isOwner = true),
@@ -569,7 +387,7 @@ class RoomInfoFragment : BaseDialogFragment() {
                 ),
                 onBackClick = {},
                 onAddUsers = {},
-                onSetUserAccess = { _, _, _-> },
+                onSetUserAccess = { _, _, _ -> },
                 onGeneralLinkCreate = {},
                 onAdditionalLinkCreate = {},
                 onLinkClick = {}
