+++ app_manager/appmanager/src/main/java/app/editors/manager/ui/fragments/share/link/ShareUsersList.kt
@@ -1,0 +1,130 @@
+package app.editors.manager.ui.fragments.share.link
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.material.Icon
+import androidx.compose.material.IconButton
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.draw.rotate
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.res.vectorResource
+import androidx.compose.ui.unit.dp
+import app.documents.core.network.common.contracts.ApiContract
+import app.documents.core.network.share.models.Share
+import app.editors.manager.managers.utils.RoomUtils
+import com.bumptech.glide.integration.compose.ExperimentalGlideComposeApi
+import com.bumptech.glide.integration.compose.GlideImage
+import com.bumptech.glide.integration.compose.placeholder
+import lib.compose.ui.addIf
+import lib.compose.ui.theme.colorTextSecondary
+import lib.compose.ui.theme.colorTextTertiary
+import lib.compose.ui.views.AnimatedVisibilityVerticalFade
+import lib.compose.ui.views.AppDivider
+import lib.toolkit.base.R
+
+
+@Composable
+internal fun ShareUsersList(title: Int, portal: String?, shareList: List<Share>?, onClick: (String, Int, Boolean) -> Unit) {
+    var visible by remember { mutableStateOf(true) }
+
+    if (!shareList.isNullOrEmpty()) {
+        Column {
+            Row(modifier = Modifier.padding(start = 16.dp, end = 16.dp, top = 16.dp, bottom = 8.dp)) {
+                Text(
+                    modifier = Modifier
+                        .weight(1f)
+                        .align(Alignment.Bottom),
+                    text = stringResource(id = title, shareList.size),
+                    style = MaterialTheme.typography.subtitle2,
+                    color = MaterialTheme.colors.colorTextSecondary
+                )
+                IconButton(
+                    modifier = Modifier.size(24.dp),
+                    onClick = { visible = !visible }
+                ) {
+                    Icon(
+                        modifier = Modifier.rotate(if (visible) 0f else 180f),
+                        imageVector = ImageVector.vectorResource(R.drawable.ic_arrow_down),
+                        contentDescription = null,
+                        tint = MaterialTheme.colors.colorTextTertiary
+                    )
+                }
+            }
+            AnimatedVisibilityVerticalFade(visible = visible) {
+                Column {
+                    shareList.forEach { share ->
+                        ShareUserItem(share = share, portal = portal) {
+                            onClick.invoke(
+                                share.sharedTo.id,
+                                share.intAccess,
+                                share.sharedTo.isOwner || share.sharedTo.isAdmin
+                            )
+                        }
+                    }
+                }
+            }
+        }
+    }
+}
+
+@OptIn(ExperimentalGlideComposeApi::class)
+@Composable
+private fun ShareUserItem(share: Share, portal: String?, onClick: () -> Unit) {
+    Column(modifier = Modifier.addIf(share.canEditAccess) { clickable(onClick = onClick) }) {
+        Row(
+            modifier = Modifier
+                .padding(horizontal = 16.dp)
+                .fillMaxWidth()
+                .height(56.dp),
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+            GlideImage(
+                modifier = Modifier
+                    .padding(end = 16.dp)
+                    .clip(CircleShape)
+                    .size(40.dp)
+                    .background(MaterialTheme.colors.colorTextTertiary),
+                model = "${ApiContract.SCHEME_HTTPS}$portal${share.sharedTo.avatarMedium}",
+                loading = placeholder(app.editors.manager.R.drawable.ic_account_placeholder),
+                contentDescription = null
+            )
+            Text(
+                modifier = Modifier.weight(1f),
+                style = MaterialTheme.typography.body1,
+                text = share.sharedTo.displayName
+            )
+            Text(
+                text = stringResource(id = RoomUtils.getAccessTitleOrOwner(share)),
+                style = MaterialTheme.typography.body2,
+                color = if (share.canEditAccess)
+                    MaterialTheme.colors.colorTextSecondary else
+                    MaterialTheme.colors.colorTextTertiary
+            )
+            if (share.canEditAccess) {
+                Icon(
+                    imageVector = ImageVector.vectorResource(R.drawable.ic_arrow_right),
+                    contentDescription = null,
+                    tint = MaterialTheme.colors.colorTextTertiary
+                )
+            }
+        }
+        AppDivider(startIndent = 16.dp + 40.dp + 16.dp)
+    }
+}
