+++ app_manager/appmanager/src/main/res/values/translatable.xml
@@ -369,7 +369,6 @@
     <string name="placeholder_no_users_groups">No users and groups</string>
     <string name="placeholder_media_error">Unable to load media</string>
     <string name="placeholder_try_again">Try again</string>
-    <string name="placeholder_loading">Loading…</string>
     <string name="placeholder_loading_files">Loading files…</string>
     <string name="placeholder_loading_users">Loading users…</string>
     <string name="placeholder_loading_groups">Loading groups…</string>
